+++ smapy/actions/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from smapy.actions.base import BaseAction
+from smapy.actions.hello import World
+
+__all__ = ('BaseAction', 'World')
+
