+++ smapy/resources/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from smapy.resources.base import BaseResource
 from smapy.resources import misc
 
-__all__ = ('misc', )
+__all__ = ('BaseResource', 'misc', )
