+++ smapy/resources/base.py
