+++ tests/actions/test_base.py
@@ -3,7 +3,7 @@
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from smapy.action import BaseAction
+from smapy.actions import BaseAction
 
 
 class TestBaseAction(TestCase):
@@ -70,14 +70,14 @@ class TestBaseAction(TestCase):
         message = {'a': 'message'}
         test_action.run_local(message)
 
-        project_dir = __file__.replace('tests/test_action.py', '')
+        project_dir = __file__.replace('tests/actions/test_base.py', '')
 
         # Asserts
         exception = [
             'Traceback (most recent call last):\n',
-            '  File "{}smapy/action.py", line 63, in run_local\n'
+            '  File "{}smapy/actions/base.py", line 63, in run_local\n'
             '    self.process(message)\n'.format(project_dir),
-            '  File "{}tests/test_action.py", line 55, in process\n'
+            '  File "{}tests/actions/test_base.py", line 55, in process\n'
             '    raise Exception("An Exception")\n'.format(project_dir),
             'Exception: An Exception\n'
         ]
@@ -118,14 +118,14 @@ class TestBaseAction(TestCase):
         message = {'a': 'message'}
         self.assertRaises(SystemExit, test_action.run_local, message)
 
-        project_dir = __file__.replace('tests/test_action.py', '')
+        project_dir = __file__.replace('tests/actions/test_base.py', '')
 
         # Asserts
         exception = [
             'Traceback (most recent call last):\n',
-            '  File "{}smapy/action.py", line 63, in run_local\n'
+            '  File "{}smapy/actions/base.py", line 63, in run_local\n'
             '    self.process(message)\n'.format(project_dir),
-            '  File "{}tests/test_action.py", line 103, in process\n'
+            '  File "{}tests/actions/test_base.py", line 103, in process\n'
             '    raise SystemExit()\n'.format(project_dir),
             'SystemExit\n'
         ]
